Move remaining logic into Transport class
Also split NodeExplanation into its own class
