+++ core/src/main/java/org/elasticsearch/action/admin/cluster/allocation/ClusterAllocationExplanation.java
@@ -54,33 +54,17 @@ public final class ClusterAllocationExplanation implements ToXContent, Writeable
     private final boolean primary;
     private final String assignedNodeId;
     private final UnassignedInfo unassignedInfo;
-    private final long remainingDelayNanos;
-    private final Set<String> activeAllocationIds;
+    private final long remainingDelayMillis;
     private final Map<DiscoveryNode, NodeExplanation> nodeExplanations;
 
-    public ClusterAllocationExplanation(ShardId shard, boolean primary, @Nullable String assignedNodeId,
-                                        UnassignedInfo unassignedInfo, Map<DiscoveryNode, Decision> nodeToDecision,
-                                        Map<DiscoveryNode, Float> nodeToWeight, long remainingDelayNanos,
-                                        List<IndicesShardStoresResponse.StoreStatus> shardStores, Set<String> activeAllocationIds) {
+    public ClusterAllocationExplanation(ShardId shard, boolean primary, @Nullable String assignedNodeId, long remainingDelayMillis,
+                                        UnassignedInfo unassignedInfo, Map<DiscoveryNode, NodeExplanation> nodeExplanations) {
         this.shard = shard;
         this.primary = primary;
         this.assignedNodeId = assignedNodeId;
         this.unassignedInfo = unassignedInfo;
-        this.remainingDelayNanos = remainingDelayNanos;
-        this.activeAllocationIds = activeAllocationIds;
-        final Map<DiscoveryNode, Decision> nodeDecisions = nodeToDecision == null ? Collections.emptyMap() : nodeToDecision;
-        final Map<DiscoveryNode, Float> nodeWeights = nodeToWeight == null ? Collections.emptyMap() : nodeToWeight;
-        assert nodeDecisions.size() == nodeWeights.size() : "decision and weight list should be the same size";
-        final Map<DiscoveryNode, IndicesShardStoresResponse.StoreStatus> storeStatuses = calculateStoreStatuses(shardStores);
-
-        this.nodeExplanations = new HashMap<>(nodeDecisions.size());
-        for (Map.Entry<DiscoveryNode, Decision> entry : nodeDecisions.entrySet()) {
-            final DiscoveryNode node = entry.getKey();
-            final Decision decision = entry.getValue();
-            final NodeExplanation nodeExplanation = calculateNodeExplanation(node, decision, nodeWeights.get(node),
-                    storeStatuses.get(node), assignedNodeId, activeAllocationIds);
-            nodeExplanations.put(node, nodeExplanation);
-        }
+        this.remainingDelayMillis = remainingDelayMillis;
+        this.nodeExplanations = nodeExplanations;
     }
 
     public ClusterAllocationExplanation(StreamInput in) throws IOException {
@@ -88,14 +72,7 @@ public final class ClusterAllocationExplanation implements ToXContent, Writeable
         this.primary = in.readBoolean();
         this.assignedNodeId = in.readOptionalString();
         this.unassignedInfo = in.readOptionalWriteable(UnassignedInfo::new);
-        this.remainingDelayNanos = in.readVLong();
-
-        int allocIdSize = in.readVInt();
-        Set<String> activeIds = new HashSet<>(allocIdSize);
-        for (int i = 0; i < allocIdSize; i++) {
-            activeIds.add(in.readString());
-        }
-        this.activeAllocationIds = activeIds;
+        this.remainingDelayMillis = in.readVLong();
 
         int mapSize = in.readVInt();
         Map<DiscoveryNode, NodeExplanation> nodeToExplanation = new HashMap<>(mapSize);
@@ -112,77 +89,12 @@ public final class ClusterAllocationExplanation implements ToXContent, Writeable
         out.writeBoolean(this.isPrimary());
         out.writeOptionalString(this.getAssignedNodeId());
         out.writeOptionalWriteable(this.getUnassignedInfo());
-        out.writeVLong(remainingDelayNanos);
-        out.writeVInt(activeAllocationIds.size());
-        for (String id : activeAllocationIds) {
-            out.writeString(id);
-        }
+        out.writeVLong(remainingDelayMillis);
 
         out.writeVInt(this.nodeExplanations.size());
         for (NodeExplanation explanation : this.nodeExplanations.values()) {
             explanation.writeTo(out);
         }
-    }
-
-    private NodeExplanation calculateNodeExplanation(DiscoveryNode node,
-                                                     Decision nodeDecision,
-                                                     Float nodeWeight,
-                                                     IndicesShardStoresResponse.StoreStatus storeStatus,
-                                                     String assignedNodeId,
-                                                     Set<String> activeAllocationIds) {
-        FinalDecision finalDecision;
-        StoreCopy storeCopy;
-        String finalExplanation;
-
-        if (node.getId().equals(assignedNodeId)) {
-            finalDecision = FinalDecision.ALREADY_ASSIGNED;
-            finalExplanation = "the shard is already assigned to this node";
-        } else if (nodeDecision.type() == Decision.Type.NO) {
-            finalDecision = FinalDecision.NO;
-            finalExplanation = "the shard cannot be assigned because one or more allocation decider returns a 'NO' decision";
-        } else {
-            finalDecision = FinalDecision.YES;
-            finalExplanation = "the shard can be assigned";
-        }
-
-        if (storeStatus != null) {
-            final Throwable storeErr = storeStatus.getStoreException();
-            if (storeErr != null) {
-                finalDecision = FinalDecision.NO;
-                if (ExceptionsHelper.unwrapCause(storeErr) instanceof CorruptIndexException) {
-                    storeCopy = StoreCopy.CORRUPT;
-                    finalExplanation = "the copy of data in the shard store is corrupt";
-                } else {
-                    storeCopy = StoreCopy.IO_ERROR;
-                    finalExplanation = "there was an IO error reading from data in the shard store";
-                }
-            } else if (activeAllocationIds.isEmpty() || activeAllocationIds.contains(storeStatus.getAllocationId())) {
-                // If either we don't have allocation IDs, or they contain the store allocation id, show the allocation
-                // status
-                storeCopy = StoreCopy.AVAILABLE;
-                if (finalDecision != FinalDecision.NO) {
-                    finalExplanation = "the shard can be assigned and the node contains a valid copy of the shard data";
-                }
-            } else {
-                // Otherwise, this is a stale copy of the data (allocation ids don't match)
-                storeCopy = StoreCopy.STALE;
-                finalExplanation = "the copy of the shard is stale, allocation ids do not match";
-                finalDecision = FinalDecision.NO;
-            }
-        } else {
-            // No copies of the data, so deciders are what influence the decision and explanation
-            storeCopy = StoreCopy.NONE;
-        }
-        return new NodeExplanation(node, nodeDecision, nodeWeight, storeStatus, finalDecision, finalExplanation, storeCopy);
-    }
-
-    private static Map<DiscoveryNode, IndicesShardStoresResponse.StoreStatus> calculateStoreStatuses(
-            List<IndicesShardStoresResponse.StoreStatus> shardStores) {
-        Map<DiscoveryNode, IndicesShardStoresResponse.StoreStatus> nodeToStatus = new HashMap<>(shardStores.size());
-        for (IndicesShardStoresResponse.StoreStatus status : shardStores) {
-            nodeToStatus.put(status.getNode(), status);
-        }
-        return nodeToStatus;
     }
 
     /** Return the shard that the explanation is about */
@@ -213,13 +125,8 @@ public final class ClusterAllocationExplanation implements ToXContent, Writeable
     }
 
     /** Return the remaining allocation delay for this shard in nanoseconds */
-    public long getRemainingDelayNanos() {
-        return this.remainingDelayNanos;
-    }
-
-    /** Return a set of the active allocation ids for this shard */
-    public Set<String> getActiveAllocationIds() {
-        return this.activeAllocationIds;
+    public long getRemainingDelayMillis() {
+        return this.remainingDelayMillis;
     }
 
     /** Return a map of node to the explanation for that node */
@@ -245,10 +152,8 @@ public final class ClusterAllocationExplanation implements ToXContent, Writeable
             if (unassignedInfo != null) {
                 unassignedInfo.toXContent(builder, params);
                 long delay = unassignedInfo.getLastComputedLeftDelayNanos();
-                builder.field("allocation_delay", TimeValue.timeValueNanos(delay));
-                builder.field("allocation_delay_ms", TimeValue.timeValueNanos(delay).millis());
-                builder.field("remaining_delay", TimeValue.timeValueNanos(remainingDelayNanos));
-                builder.field("remaining_delay_ms", TimeValue.timeValueNanos(remainingDelayNanos).millis());
+                builder.timeValueField("allocation_delay_ms", "allocation_delay", TimeValue.timeValueNanos(delay));
+                builder.timeValueField("remaining_delay_ms", "remaining_delay", TimeValue.timeValueMillis(remainingDelayMillis));
             }
             builder.startObject("nodes");
             for (NodeExplanation explanation : nodeExplanations.values()) {
@@ -258,116 +163,6 @@ public final class ClusterAllocationExplanation implements ToXContent, Writeable
         }
         builder.endObject(); // end wrapping object
         return builder;
-    }
-
-    /** The cluster allocation explanation for a single node */
-    public class NodeExplanation implements Writeable, ToXContent {
-        private final DiscoveryNode node;
-        private final Decision nodeDecision;
-        private final Float nodeWeight;
-        private final IndicesShardStoresResponse.StoreStatus storeStatus;
-        private final FinalDecision finalDecision;
-        private final String finalExplanation;
-        private final StoreCopy storeCopy;
-
-        public NodeExplanation(final DiscoveryNode node, final Decision nodeDecision, final Float nodeWeight,
-                               final @Nullable IndicesShardStoresResponse.StoreStatus storeStatus,
-                               final FinalDecision finalDecision, final String finalExplanation, final StoreCopy storeCopy) {
-            this.node = node;
-            this.nodeDecision = nodeDecision;
-            this.nodeWeight = nodeWeight;
-            this.storeStatus = storeStatus;
-            this.finalDecision = finalDecision;
-            this.finalExplanation = finalExplanation;
-            this.storeCopy = storeCopy;
-        }
-
-        public NodeExplanation(StreamInput in) throws IOException {
-            this.node = new DiscoveryNode(in);
-            this.nodeDecision = Decision.readFrom(in);
-            this.nodeWeight = in.readFloat();
-            if (in.readBoolean()) {
-                this.storeStatus = IndicesShardStoresResponse.StoreStatus.readStoreStatus(in);
-            } else {
-                this.storeStatus = null;
-            }
-            this.finalDecision = FinalDecision.readFrom(in);
-            this.finalExplanation = in.readString();
-            this.storeCopy = StoreCopy.readFrom(in);
-        }
-
-        @Override
-        public void writeTo(StreamOutput out) throws IOException {
-            node.writeTo(out);
-            Decision.writeTo(nodeDecision, out);
-            out.writeFloat(nodeWeight);
-            if (storeStatus == null) {
-                out.writeBoolean(false);
-            } else {
-                out.writeBoolean(true);
-                storeStatus.writeTo(out);
-            }
-            finalDecision.writeTo(out);
-            out.writeString(finalExplanation);
-            storeCopy.writeTo(out);
-        }
-
-        public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
-            builder.startObject(node.getId()); {
-                builder.field("node_name", node.getName());
-                builder.startObject("node_attributes"); {
-                    for (Map.Entry<String, String> attrEntry : node.getAttributes().entrySet()) {
-                        builder.field(attrEntry.getKey(), attrEntry.getValue());
-                    }
-                }
-                builder.endObject(); // end attributes
-                builder.startObject("store"); {
-                    builder.field("shard_copy", storeCopy.toString());
-                    if (storeStatus != null) {
-                        final Throwable storeErr = storeStatus.getStoreException();
-                        if (storeErr != null) {
-                            builder.field("store_exception", ExceptionsHelper.detailedMessage(storeErr));
-                        }
-                    }
-                }
-                builder.endObject(); // end store
-                builder.field("final_decision", finalDecision.toString());
-                builder.field("final_explanation", finalExplanation.toString());
-                builder.field("weight", nodeWeight);
-                nodeDecision.toXContent(builder, params);
-            }
-            builder.endObject(); // end node <uuid>
-            return builder;
-        }
-
-        public DiscoveryNode getNode() {
-            return this.node;
-        }
-
-        public Decision getDecision() {
-            return this.nodeDecision;
-        }
-
-        public Float getWeight() {
-            return this.nodeWeight;
-        }
-
-        @Nullable
-        public IndicesShardStoresResponse.StoreStatus getStoreStatus() {
-            return this.storeStatus;
-        }
-
-        public FinalDecision getFinalDecision() {
-            return this.finalDecision;
-        }
-
-        public String getFinalExplanation() {
-            return this.finalExplanation;
-        }
-
-        public StoreCopy getStoreCopy() {
-            return this.storeCopy;
-        }
     }
 
     /** An Enum representing the final decision for a shard allocation on a node */
@@ -426,7 +221,9 @@ public final class ClusterAllocationExplanation implements ToXContent, Writeable
         // There was an error reading this node's copy of the data
         IO_ERROR((byte) 3),
         // The copy of the data on the node is stale
-        STALE((byte) 4);
+        STALE((byte) 4),
+        // It's unknown what the copy of the data is
+        UNKNOWN((byte) 5);
 
         private final byte id;
 
@@ -441,6 +238,7 @@ public final class ClusterAllocationExplanation implements ToXContent, Writeable
                 case 2: return CORRUPT;
                 case 3: return IO_ERROR;
                 case 4: return STALE;
+                case 5: return UNKNOWN;
                 default:
                     throw new IllegalArgumentException("unknown id for store copy: [" + id + "]");
             }
@@ -454,6 +252,7 @@ public final class ClusterAllocationExplanation implements ToXContent, Writeable
                 case 2: return "CORRUPT";
                 case 3: return "IO_ERROR";
                 case 4: return "STALE";
+                case 5: return "UNKNOWN";
                 default:
                     throw new IllegalArgumentException("unknown id for store copy: [" + id + "]");
             }
